feat: make the goofys CLI embeddable and fix macOS builds (#3)
