+++ .github/workflows/release.yml
@@ -22,8 +22,8 @@ jobs:
         env:
           CGO_ENABLED: '0'
         run: |
-          GOOS=linux GOARCH=amd64 go build -ldflags "-s -w" -o goofys-amd64 .
-          GOOS=linux GOARCH=arm64 go build -ldflags "-s -w" -o goofys-arm64 .
+          GOOS=linux GOARCH=amd64 go build -ldflags "-s -w -X main.Version=$GITHUB_REF_NAME" -o goofys-amd64 .
+          GOOS=linux GOARCH=arm64 go build -ldflags "-s -w -X main.Version=$GITHUB_REF_NAME" -o goofys-arm64 .
           cp goofys-amd64 goofys
 
       - name: Create release and upload assets
